+++ include/tudat/astro/orbit_determination/observation_partials/directObservationPartial.h
@@ -166,7 +166,7 @@ public:
      */
     int getNumberOfLighTimeCorrectionPartialsFunctions( )
     {
-        return lighTimeCorrectionPartialsFunctions_.size( );
+        return static_cast< int >( lighTimeCorrectionPartialsFunctions_.size( ) );
     }
 
     virtual bool useLinkIndependentPartials( )
